+++ docs/mobile-release.md
@@ -198,7 +198,12 @@ and a native incompatibility must be corrected with a higher-version signed APK/
 
 ### Publish OTA updates from GitHub Actions
 
-The `Publish Expo OTA Update` workflow is manual-only and publishes Android phone JavaScript/assets from `master`.
+The `Publish Expo OTA Update` workflow runs whenever a commit lands on `master`. It publishes Android phone
+JavaScript/assets to the `internal` channel first, then starts a dependent production job. The production job targets
+the protected GitHub `production` environment, so it remains blocked until a required reviewer verifies the internal
+devices and approves the deployment. A newer `master` commit cancels an older pending promotion so stale code cannot
+be approved after a newer internal update.
+
 Before its first use:
 
 1. Create an Expo access token and save it as the repository Actions secret `EXPO_TOKEN`.
@@ -208,18 +213,17 @@ Before its first use:
    `EXPO_UPDATES_CHANNEL=production`.
 4. Build and install the phone binary from a known `master` commit with the same channel/environment you intend to
    update.
+5. Keep required reviewers and protected-branch deployment rules enabled on the GitHub `production` environment.
 
-In GitHub, open **Actions > Publish Expo OTA Update > Run workflow**, select `master`, and provide:
+The workflow pulls and validates each selected EAS environment, then publishes the merge commit with its commit
+message. No native build tag is supplied. Expo selects clients by project, Android platform, channel, and the
+`appVersion` runtime generated by `mobile/app.config.js`; clients on an older or otherwise different runtime do not
+receive the update. Increment the mobile app version and create a new signed native build whenever native
+dependencies, config plugins, icons, permissions, or other native inputs change. OTA updates never update the Wear
+app.
 
-- `native_build_ref`: the native build tag shown in **Settings > About Calibrate**, or the exact source commit used to
-  build the installed phone app.
-- `channel`: `internal` for the dogfood APK or `production` for a production-channel build.
-- `message`: a short description shown in EAS Update history.
-
-The workflow pulls and validates the selected EAS environment, verifies that the current source descends from the
-native build reference, and compares the precise runtime fingerprint before it uploads. Any app version, native
-dependency, config plugin, icon, Wear source, or other native input change stops the publish and requires a new signed
-phone/Watch build. OTA updates never update the Wear app.
+After testing the automatic internal deployment, open its workflow run in GitHub Actions and approve the waiting
+`production` environment deployment. Reject it or leave it pending if verification does not pass.
 
 ## Preserve on-device data during upgrades
